# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Swagger documentation untuk layanan CatatUangku. Import file JSON-nya ke Postman untuk koleksi otomatis.",
        "title": "CatatUangku API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
...
        "/invitations": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Invitations"
                ],
                "summary": "Kirim undangan group",
                "parameters": [
                    {
                        "description": "Data undangan",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/invitations.InvitationRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/handler.InvitationDataResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/invitations/accept": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Invitations"
                ],
                "summary": "Terima undangan group",
                "parameters": [
                    {
                        "description": "Data penerimaan undangan",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/invitations.AcceptInvitationRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.InvitationDataResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/transactions": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Transactions"
                ],
                "summary": "Daftar transaksi",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.TransactionListResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Transactions"
                ],
                "summary": "Buat transaksi",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Group ID (isi saat scope=group)",
                        "name": "group_id",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Category ID",
                        "name": "category_id",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "User ID pembuat",
                        "name": "created_by_user_id",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Account ID",
                        "name": "account_id",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "income atau expense",
                        "name": "type",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "number",
                        "description": "Nominal transaksi",
                        "name": "total_amount",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "personal atau group",
                        "name": "scope",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Catatan",
                        "name": "description",
                        "in": "formData"
                    },
                    {
                        "type": "file",
                        "description": "Lampiran transaksi (boleh multiple)",
                        "name": "attachments",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/handler.TransactionDataResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/transactions/account/{account_id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Transactions"
                ],
                "summary": "Transaksi berdasarkan akun",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Account ID",
                        "name": "account_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.TransactionListResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/transactions/user/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Transactions"
                ],
                "summary": "Transaksi berdasarkan user",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.TransactionListResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/transactions/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Transactions"
                ],
                "summary": "Detail transaksi",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Transaction ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.TransactionResponseBody"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "attachments.Attachment": {
            "type": "object",
            "properties": {
                "file_name": {
                    "type": "string"
                },
                "file_path": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "transaction_id": {
                    "type": "string"
                },
                "uploaded_at": {
                    "type": "string"
                }
            }
        },
        "attachments.AttachmentResponse": {
            "type": "object",
            "properties": {
                "file_name": {
                    "type": "string"
                },
                "file_url": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "transaction_id": {
                    "type": "string"
                },
                "uploaded_at": {
                    "type": "string"
                }
            }
        },
...
        "handler.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
...
        "handler.InvitationDataResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/invitations.Invitation"
                }
            }
        },
...
        "handler.TransactionDataResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/transactions.Transactions"
                }
            }
        },
        "handler.TransactionListResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/transactions.TransactionResponse"
                    }
                }
            }
        },
        "handler.TransactionResponseBody": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/transactions.TransactionResponse"
                }
            }
        },
...
        "invitations.AcceptInvitationRequest": {
            "type": "object",
            "required": [
                "group_id",
                "invitee_email",
                "token",
                "user_id"
            ],
            "properties": {
                "group_id": {
                    "type": "string"
                },
                "invitee_email": {
                    "type": "string"
                },
                "token": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "invitations.Invitation": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "expired_at": {
                    "type": "string"
                },
                "group_id": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "invitee_email": {
                    "type": "string"
                },
                "inviter_user_id": {
                    "type": "string"
                },
                "role": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "token": {
                    "type": "string"
                }
            }
        },
        "invitations.InvitationRequest": {
            "type": "object",
            "required": [
                "group_id",
                "invitee_email",
                "inviter_user_id",
                "role"
            ],
            "properties": {
                "group_id": {
                    "type": "string"
                },
                "invitee_email": {
                    "type": "string"
                },
                "inviter_user_id": {
                    "type": "string"
                },
                "role": {
                    "description": "Token\t\t\t string `json:\"token\"`",
                    "type": "string"
                }
            }
        },
        "transactionlines.TransactionLine": {
            "type": "object",
            "properties": {
                "account_id": {
                    "type": "string"
                },
                "credit": {
                    "type": "number"
                },
                "debit": {
                    "type": "number"
                },
                "id": {
                    "type": "string"
                },
                "note": {
                    "type": "string"
                },
                "transaction_id": {
                    "type": "string"
                }
            }
        },
        "transactions.TransactionResponse": {
            "type": "object",
            "properties": {
                "attachments": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/attachments.AttachmentResponse"
                    }
                },
                "category_id": {
                    "type": "string"
                },
                "created_by_user_id": {
                    "type": "string"
                },
                "date": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "group_id": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "scope": {
                    "type": "string"
                },
                "total_amount": {
                    "type": "number"
                },
                "transaction_lines": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/transactionlines.TransactionLine"
                    }
                },
                "type": {
                    "type": "string"
                }
            }
        },
        "transactions.Transactions": {
            "type": "object",
            "properties": {
                "attachments": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/attachments.Attachment"
                    }
                },
                "category_id": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "created_by_user_id": {
                    "type": "string"
                },
                "date": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "group_id": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "private_to_user_id": {
                    "type": "string"
                },
                "scope": {
                    "type": "string"
                },
                "total_amount": {
                    "type": "number"
                },
                "transaction_lines": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/transactionlines.TransactionLine"
                    }
                },
                "type": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}